requirements.txt
asgiref==3.7.2
asttokens==3.0.0
branca==0.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
Django==4.2.7
django-bootstrap5==23.3
django-cors-headers==4.7.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
executing==2.2.0
folium==0.19.5
idna==3.10
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.5
openpyxl==3.1.2
packaging==25.0
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.7
polyline==2.0.2
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg==3.2.7
psycopg-binary==3.2.7
psycopg2-binary==2.9.9
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pywin32==310
pyzmq==26.4.0
requests==2.32.3
six==1.17.0
sqlparse==0.4.4
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.4.0
wcwidth==0.2.13
xyzservices==2025.1.0
dj-database-url==2.1.0
gunicorn==21.2.0
pytz==2023.3.post1
whitenoise==6.6.0
